Move SdkLogListener & SdkLogInitializer(->LogHelper) to Kotlin package
